Introduce `WP_Fail2Ban_Redux::user_enumeration`
This function will replace `WP_Fail2Ban_Redux::redirect_canonical` in a future commit. This is being done to avoid the extra overhead inccured by the `redirect_canonical` function. Since Fail2Ban is used to help mitigate brute-force logins and DDoS attacks (both of which waste server resources), it's a good idea to do as little processing as possible before ending a malicious request. The `redirect_canonical` filter was not the place for this.

Instead, we now hook into the `parse_request` action. While, the `init` action would be even better, hooking into `parse_request`, with a priority of 12, is a cheap way to avoid blocking WP-API requests.

The following changes were also made, that differ from the previous behavior of `WP_Fail2Ban_Redux::redirect_canonical`:

* We now check for both the `author` and `author_name` parameters of a GET request. Otherwise, we bail. While the `author_name` parameter doesn't produce the exact same behaviors as the `author` parameter, it is a valid method to verify the existence of a user. This knowledge could then be used to mount a more targeted brute-force login attack.
* We now check make sure we're not in the admin. Fixes #2.
* We check to ensure that permalinks are enabled. Otherwise, we bail, since user enumeration is essentially required in this scenario.