Bail early from `WP_Fail2Ban_Redux::authenticate` if no username passed
Prior to this commit, filtering `wp_fail2ban_redux_blocked_users_not_in` to `true` would cause a failed login attempt to be logged on logout, or when simply visiting `wp-login.php`. This is incorrect behavior. Returning early is the correct behavior, as there's no reason to run a blocked username check, when no username is passed. So, fixing that, also fixes the logout issue.